feat: Refactor network switching functionality into a hook.
